confluence_system/sierra_chart/requirements.txt
# Sierra Chart Integration Requirements
supabase==2.5.0
python-dotenv==1.0.0
python-dateutil==2.8.2